Support systemd socket activation in `zb serve`
Fixes #121

+++ cmd/zb/serve.go
@@ -20,6 +20,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/coreos/go-systemd/v22/activation"
 	"github.com/spf13/cobra"
 	"golang.org/x/sync/errgroup"
 	"zb.256lights.llc/pkg/internal/backend"
@@ -45,6 +46,7 @@ type serveOptions struct {
 	allowKeepFailed   bool
 	coresPerBuild     int
 	buildLogRetention time.Duration
+	systemdSocket     bool
 
 	webListenAddress   string
 	allowRemoteWeb     bool
@@ -68,6 +70,9 @@ func newServeCommand(g *globalConfig) *cobra.Command {
 	}
 	if osutil.IsRoot() {
 		opts.buildUsersGroup = backend.DefaultBuildUsersGroup
+	}
+	if runtime.GOOS == "linux" {
+		c.Flags().BoolVar(&opts.systemdSocket, "systemd", false, "use systemd socket activation")
 	}
 	c.Flags().StringVar(&opts.dbPath, "db", opts.dbPath, "`path` to store database file")
 	c.Flags().StringVar(&opts.buildDir, "build-root", os.TempDir(), "`dir`ectory to store temporary build artifacts")
@@ -133,19 +138,32 @@ func runServe(ctx context.Context, g *globalConfig, opts *serveOptions) error {
 		webHandler.staticAssets = ui.StaticAssets()
 	}
 
-	l, err := listenUnix(g.storeSocket)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if err := os.Remove(g.storeSocket); err != nil && !errors.Is(err, os.ErrNotExist) {
-			log.Warnf(ctx, "Failed to clean up socket: %v", err)
+	var l net.Listener
+	if runtime.GOOS == "linux" && opts.systemdSocket {
+		listeners, err := activation.Listeners()
+		if err != nil {
+			return err
 		}
-	}()
+		if len(listeners) != 1 {
+			return fmt.Errorf("systemd passed in %d sockets (want 1)", len(listeners))
+		}
+		l = listeners[0]
+	} else {
+		var err error
+		l, err = listenUnix(g.storeSocket)
+		if err != nil {
+			return err
+		}
+		defer func() {
+			if err := os.Remove(g.storeSocket); err != nil && !errors.Is(err, os.ErrNotExist) {
+				log.Warnf(ctx, "Failed to clean up socket: %v", err)
+			}
+		}()
+	}
 
 	grp, grpCtx := errgroup.WithContext(ctx)
 
-	openConns := make(sets.Set[*net.UnixConn])
+	openConns := make(sets.Set[net.Conn])
 	var openConnsMu sync.Mutex
 	grp.Go(func() error {
 		// Once the context is Done, refuse new connections and RPCs.
@@ -157,7 +175,7 @@ func runServe(ctx context.Context, g *globalConfig, opts *serveOptions) error {
 		}
 		openConnsMu.Lock()
 		for conn := range openConns.All() {
-			if err := conn.CloseRead(); err != nil {
+			if err := closeRead(conn); err != nil {
 				log.Errorf(grpCtx, "Closing Unix socket: %v", err)
 			}
 		}
@@ -184,7 +202,7 @@ func runServe(ctx context.Context, g *globalConfig, opts *serveOptions) error {
 
 	grp.Go(func() error {
 		for {
-			conn, err := l.AcceptUnix()
+			conn, err := l.Accept()
 			if errors.Is(err, net.ErrClosed) {
 				return nil
 			}
@@ -342,6 +360,14 @@ func listenUnix(path string) (*net.UnixListener, error) {
 	}
 
 	return l, nil
+}
+
+func closeRead(c net.Conn) error {
+	cr, ok := c.(interface{ CloseRead() error })
+	if !ok {
+		return fmt.Errorf("%T does not support uni-directional close", c)
+	}
+	return cr.CloseRead()
 }
 
 type pathMapFlag map[string]string
